Sanitize exchange paths and improve matching
The sanitizing via regex-matching simplifies exception cases with the
wrapping ensuring that invalid exchange keys will not lead to a crash
as occured in testing originally on pubmatic/1.

The optional argument matcher means the matcher's definition should
normally only be evaluated once, although even if it is re-evaluated,
the regex engine memoizes compilations.